Move notein into separate "midi" module
