+++ lib/glue.js
@@ -1167,42 +1167,6 @@ exports.declareObjects = function(library) {
 
   })
 
-  library['notein'] = PdObject.extend({
-
-    type: 'notein',
-
-    inletDefs: [],
-
-    outletDefs: [portlets.Outlet, portlets.Outlet, portlets.Outlet],
-
-    init: function(args) {
-      this._eventReceiver = new mixins.EventReceiver()
-      this._eventReceiver.on(pdGlob.emitter, 'midiMessage', this._onMidiMessage.bind(this))
-      this._channel = args[0]
-    },
-
-    _onMidiMessage: function(midiMessage) {
-      var data = midiMessage.data
-        , event = data[0] >> 4
-      // Respond to note on and note off events
-      if (event === 8 || event === 9) {
-        var channel = (data[0] & 0x0F) + 1
-          , note = data[1]
-          , velocity = data[2]
-        if (typeof this._channel === 'number') {
-          if (channel === this._channel) {
-            this.o(0).message([note])
-            this.o(1).message([velocity])
-          }
-        } else {
-          this.o(0).message([note])
-          this.o(1).message([velocity])   
-          this.o(2).message([channel])
-        }
-      }
-    }
-  })
-
   library['pd'] = library['graph'] = Patch
 
 }
