+++ lib/index.js
@@ -24,4 +24,5 @@ exports.declareObjects = function(library) {
   require('./controls').declareObjects(library)
   require('./waa/dsp').declareObjects(library)
   require('./waa/portlets').declareObjects(library)
+  require('./midi').declareObjects(library)
 }
